feat(viewer): various fixes, fiddles
